--- src/applications/financial-status-report/pages/income/spouse/additionalIncome.js
@@ -1,80 +1,0 @@
-import ItemLoop from '../../../components/ItemLoop';
-import TableDetailsView from '../../../components/TableDetailsView';
-import currencyUI from 'platform/forms-system/src/js/definitions/currency';
-import Typeahead from '../../../components/Typeahead';
-import {
-  formatOptions,
-  incomeTypes,
-} from '../../../constants/typeaheadOptions';
-import _ from 'lodash/fp';
-
-export const uiSchema = {
-  'ui:title': 'Your spouse information',
-  spouseAdditionalIncome: {
-    'ui:title': 'Does your spouse currently receive any additional income?',
-    'ui:widget': 'yesNo',
-    'ui:required': () => true,
-  },
-  hasAdditionalIncome: {
-    'ui:options': {
-      expandUnder: 'spouseAdditionalIncome',
-    },
-    spouseAdditionalIncome: {
-      'ui:description':
-        'Please provide information about additional income your spouse currently receives.',
-      'ui:field': ItemLoop,
-      'ui:options': {
-        viewType: 'table',
-        viewField: TableDetailsView,
-        doNotScroll: true,
-        showSave: true,
-        itemName: 'income',
-      },
-      items: {
-        incomeType: {
-          'ui:title': 'Type of income',
-          'ui:field': Typeahead,
-          'ui:options': {
-            classNames: 'input-size-3',
-            getOptions: () => formatOptions(incomeTypes),
-          },
-          'ui:required': formData => formData.spouseAdditionalIncome,
-        },
-        incomeAmount: _.merge(currencyUI('Monthly income amount'), {
-          'ui:options': {
-            widgetClassNames: 'input-size-2',
-          },
-          'ui:required': formData => formData.spouseAdditionalIncome,
-        }),
-      },
-    },
-  },
-};
-export const schema = {
-  type: 'object',
-  properties: {
-    spouseAdditionalIncome: {
-      type: 'boolean',
-    },
-    hasAdditionalIncome: {
-      type: 'object',
-      properties: {
-        spouseAdditionalIncome: {
-          type: 'array',
-          items: {
-            type: 'object',
-            required: ['incomeType', 'incomeAmount'],
-            properties: {
-              incomeType: {
-                type: 'string',
-              },
-              incomeAmount: {
-                type: 'number',
-              },
-            },
-          },
-        },
-      },
-    },
-  },
-};
